Model sampling implemented but buggy.
One can successfully sample a model, but it seems there's some issue with the randomness when producing a block; despite seeding it, the blocks don't all come out with the same connections.

Loading/storing weights is also essentially implemented - it's not tested, but I think it should work as long as the proper names are used in the model. This isn't the case yet, so there's some fixing to be done with `variable_scope`s (and perhaps a little name manipulation in the weight loader as well).

Mutations aren't supported at all yet, neither is actually training a model and evaluating its fitness.