Everything except mutations tested + working (probably).
evo was renamed to architecture.py to be a bit more descriptive; the name of the module was also changed to nn_evolution.

Each model is now built + trained in a separate thread so that there aren't issues about having multiple computational graphs. This isolation seems to be working well but does mean that the `Architecture` doesn't have access to it's `keras` model anymore. When training is done, the fitness is saved but the model itself is lost.

Serialization / deserialization of `Architecture`s is implemented; this does still need a bit of testing. The arches are saved as soon as their fitness is known, so we should be able to restart an experiment with the same population it had when it ended/was killed.

The fitness is the correlation (Pearson) with the labels now instead of the MSE (or, more generally, the loss). This is hard-coded and would need to be made more flexible if using labels where this doesn't make sense (e.g. classification).

Mutation is implemented, but I think there may be errors still. Rather than having a specified number of mutations (e.g. change one op or input of one node in a block), there are mutation probabilities for whether each node is mutated and for whether the ops/inputs are mutated within a node. It should be easy to implement different mutation rates at different points during an experiment if desired (e.g. more mutations at the start).

`logging` is used in lieu of `print`; the messages that previously only showed up under `verbose` now always show up. `verbose` controls whether debug messages also show (so far these just document building an arch out of blocks).

`main.py` actually works now (up to issues with mutations and possible `Architecture` deserialization): data loading is there (hard-coded to be fore the sequences; needs more flexibility sometime), Comet is used if a key is found (with basic metrics - fitness of the newest model, average fitness), and saving/loading of `Architecture`s as mentioned above.

`run.sh` can be edited and used to more easily enter all of the command-line arguments to run `main.py`.