Fixed `Architecture` (de)serialization.
Arches are serialized now using the signature of `Architecture` and `Block`, so even if new arguments are added, no code changes should be required in serialization.

If an arch has already been trained (has a fitness), it isn't trained again in `main.py`.

If there's an error during training, `main.py` asserts `False` (because the error will have been in a separate thread) and saves the error-causing arch for later inspection.