Colored debug messages, fix error in mutation.
The messages that log when a repeat/block/node/input is being built are now colored blue to make it easier to find the region you're looking for.

There was an error in mutating that could allow an Id op to be used when striding was needed; I've changed the boolean logic to be more straightforward and hopefully fix this.